Adding persistent collectives
