+++ src/collective/CMakeLists.txt
@@ -6,6 +6,7 @@ set(collective_HEADERS
     collective/allgather.h
     collective/gather.h
     collective/bcast.h
+    collective/alltoall_init.h
     PARENT_SCOPE
     )
 
@@ -15,6 +16,7 @@ set(collective_SOURCES
     collective/allgather.c
     collective/gather.c
     collective/bcast.c
+    collective/alltoall_init.c
     PARENT_SCOPE
     )
