+++ src/mpi_advance.h
@@ -12,9 +12,11 @@
 #include "collective/allgather.h"
 #include "collective/alltoall.h"
 #include "collective/alltoallv.h"
+#include "collective/alltoall_init.h"
 
 #include "neighborhood/dist_graph.h"
 #include "neighborhood/neighbor.h"
+#include "neighborhood/neighbor_persistent.h"
 
 #ifdef GPU
     #include "heterogeneous/gpu_alltoall.h"
