+++ src/neighborhood/CMakeLists.txt
@@ -3,6 +3,7 @@ set(CMAKE_INCLUDE_CURRENT_DIR ON)
 set(neighborhood_HEADERS
     neighborhood/dist_graph.h
     neighborhood/neighbor.h
+    neighborhood/neighbor_persistent.h
     PARENT_SCOPE
     )
 
@@ -10,6 +11,7 @@ set(neighborhood_SOURCES
     neighborhood/dist_graph.c
     neighborhood/neighbor.c
     neighborhood/neighbor_locality.cpp
+    neighborhood/neighbor_persistent.c
     PARENT_SCOPE
     )
